chore: merge the utils module
