--- mw2fcitx/utils/__init__.py
@@ -1,6 +1,0 @@
-from .dedup import dedup
-from .normalize import normalize
-from .logger import console
-from .sanitize import sanitize
-from .is_libime_used import is_libime_used
-from .smart_rewrite import smart_rewrite
