--- mw2fcitx/utils/dedup.py
@@ -1,2 +1,0 @@
-def dedup(arr: [str]):
-    return list(set(arr))
