--- mw2fcitx/utils/is_libime_used.py
@@ -1,6 +1,0 @@
-def is_libime_used(config):
-    generators = config.get('generator') or []
-    for i in generators:
-        if i.get("use") == "pinyin":
-            return True
-    return False
