--- mw2fcitx/utils/normalize.py
@@ -1,2 +1,0 @@
-def normalize(word):
-    return word.strip()
